refactor: extract displayChoicesTool into tools/displayChoices.ts
