+++ src/server/albert/Albert.ts
@@ -3,13 +3,11 @@ import {
   convertToModelMessages,
   stepCountIs,
   streamText as aiStreamText,
-  tool,
   ToolSet,
   UIMessage,
   wrapLanguageModel,
 } from "ai";
 import type { LanguageModelV3 } from "@ai-sdk/provider";
-import { z } from "zod";
 import { Prisma } from "@prisma/client";
 import { devToolsMiddleware } from "@ai-sdk/devtools";
 import { configuration } from "@/config";
@@ -21,40 +19,6 @@ export function withOptionalDevTools(model: LanguageModelV3): LanguageModelV3 {
   }
   return wrapLanguageModel({ model, middleware: devToolsMiddleware() });
 }
-
-// TODO : extraire dans son propre fichier
-export const displayChoicesInputSchema = z.object({
-  question: z
-    .string()
-    .describe("Question affichée en haut du panneau de choix"),
-  choices: z
-    .array(
-      z.object({
-        label: z.string().describe("Texte à afficher sur le bouton"),
-        value: z
-          .string()
-          .describe("Valeur à renvoyer lorsque le bouton est cliqué"),
-      }),
-    )
-    .describe("Liste des choix à proposer"),
-});
-
-export type DisplayChoice = z.infer<
-  typeof displayChoicesInputSchema
->["choices"][number];
-
-export const displayChoicesTool = tool({
-  description:
-    "Affiche des choix dans un panneau pour l'utilisateur. Le paramètre 'question' est la question affichée en haut du panneau. Utilise cet outil quand tu veux proposer des options à l'utilisateur. IMPORTANT : écris toujours ton message textuel AVANT d'appeler cet outil. Ne l'appelle jamais sans avoir d'abord rédigé le texte d'accompagnement.",
-  inputSchema: displayChoicesInputSchema,
-  execute: async ({
-    question,
-    choices,
-  }): Promise<{ question: string; choices: DisplayChoice[] }> => ({
-    question,
-    choices,
-  }),
-});
 
 const DEFAULT_MODEL = "openweight-large";
